moved models so that they are kept by proguard
