+++ Android/app/src/main/java/ru/inwords/inwords/game/presentation/game_level/GameLevelViewModel.kt
@@ -11,7 +11,6 @@ import io.reactivex.processors.PublishProcessor
 import ru.inwords.inwords.GraphGameLevelDirections
 import ru.inwords.inwords.core.resource.Resource
 import ru.inwords.inwords.core.rxjava.SchedulersFacade
-import ru.inwords.inwords.game.domain.CardsData
 import ru.inwords.inwords.game.domain.interactor.ContinueGameInteractor
 import ru.inwords.inwords.game.domain.interactor.GameInteractor
 import ru.inwords.inwords.game.domain.model.*
