+++ Android/app/src/test/java/ru/inwords/inwords/game/presentation/game_level/GameSceneTest.kt
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import ru.inwords.flipview.FlipView
 import ru.inwords.inwords.R
-import ru.inwords.inwords.game.domain.CardsData
+import ru.inwords.inwords.game.domain.model.CardsData
 import ru.inwords.inwords.translation.domain.model.WordTranslation
 import java.lang.ref.WeakReference
